♻️ refactor(lib): Separate Web Framework Integrations into frameworks/
