+++ src/frameworks/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "rocket")]
+pub mod rocket;
